#4 Renamed Rate -> Ratio
