+++ src/application/core/feature_extraction/constants.py
@@ -1,6 +1,3 @@
-dataset_path = "../data/dataset_1"
-fs = 200
-window_size = 10
 alpha_lowcut = 7.0
 alpha_highcut = 12.0
 beta_lowcut = 14.0
